Refactor frontend script into modules
