+++ static/index.html
@@ -71,6 +71,6 @@
 </div>
 
 <script src="https://cdn.jsdelivr.net/npm/canvas-confetti@1.6.0/dist/confetti.browser.min.js"></script>
-<script src="/static/script.js"></script>
+<script type="module" src="/static/js/main.js"></script>
 
 </body>
