+++ static/js/dom.js
@@ -1,0 +1,13 @@
+export const elements = {
+    form: document.getElementById("form"),
+    input: document.getElementById("word"),
+    history: document.getElementById("history"),
+    scoreboard: document.getElementById("scoreboard"),
+    roomInfo: document.getElementById("room-info"),
+    createRoomBtn: document.getElementById("create-room"),
+    joinRoomBtn: document.getElementById("join-room"),
+    roomIdInput: document.getElementById("room-id-input"),
+    playerNameInput: document.getElementById("player-name"),
+    modeSelect: document.getElementById("mode-select"),
+    messages: document.getElementById("messages"),
+};
