+++ static/js/state.js
@@ -1,0 +1,7 @@
+export const state = {
+    entries: [],
+    currentRoomId: null,
+    currentMode: "coop",
+    roomLocked: false,
+    websocket: null,
+};
